requirements.txt
Django>=5.2
psycopg2-binary
djangorestframework
django-filter
django-redis
